% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoodji.R
\name{counter}
\alias{counter}
\title{Counter Function}
\usage{
counter(text)
}
\arguments{
\item{text}{string: the input text for sentiment analysis}
}
\value{
data frame: a data frame that contains the summary statistics for character, word, and sentence count.
}
\description{
Generates a summary dataframe of the input text which contains counts for characters, words, and sentences.
}
\examples{
counter("I am very happy.")
}
