% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{A list of English words and their associations with eight basic emotions
(anger, fear, anticipation, trust, surprise, sadness, joy, and disgust)}
\format{
A data frame with 141820 rows and 3 variables:
\describe{
\item{V1}{English word, characters}
\item{V2}{Emotion, characters}
\item{V3}{Count of the word, integer}
...
}
}
\source{
\url{https://saifmohammad.com/WebPages/NRC-Emotion-Lexicon.htm}
}
\usage{
data
}
\description{
A dataset containing 141820 English words and their associations with eight basic emotions.
}
\keyword{datasets}
