% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoodji.R
\name{sentiment_df}
\alias{sentiment_df}
\title{Sentiment df Function}
\usage{
sentiment_df(text, sentiment_input = "all")
}
\arguments{
\item{text}{string: the input text for sentiment analysis}

\item{sentiment_input}{string (optional): the sentiment that the analysis focuses on, could be happy, angry, or sad etc. Defaults to "all".}
}
\value{
dataframe: a data frame that contains the summary of sentiment analysis
}
\description{
Generates a sentiment analysis summary dataframe of the input text. The summary dataframe would include
the sentiment type, sentiment words, number of sentiment words, and highest sentiment percentage.
}
