% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoodji.R
\name{sentiment_plot}
\alias{sentiment_plot}
\title{Sentiment Plot Function}
\usage{
sentiment_plot(text, sentiment)
}
\arguments{
\item{text}{factor}

\item{sentiment}{factor}
}
\value{
factor
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
