% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoodji.R
\name{textsentiment_to_emoji}
\alias{textsentiment_to_emoji}
\title{Emoji Function}
\usage{
textsentiment_to_emoji(text, sentiment_dataframe = NULL)
}
\arguments{
\item{text}{string: A text string containing english words}

\item{sentiment_dataframe}{data frame: A dataframe which contains word and key column which shows the sentiment of each word. Only supports
the Happy, Sad, Suprise, Fear and Angry as keys. If no dataframe is given the results of sentiment_df
function would be used}
}
\value{
string:  A string containing only emoji's with no words. The emojis are written in the CLDR short name format.
}
\description{
Detect the word sentiments of a text and replace the with the matching emojis.
}
